Rework reading and accessing values
Split the Status::Access enum value into Reading and Writing for
more specific status information. Store previous reads in a
separate field so that it isn't wiped out by subsequent operations
that might use the buffer and not actually do a read. It's a more
permanent place for a read value since the SPI reads are
asynchronous.

Slow down the SPI clock. I will experiment with this later to see
if this was affecting values, but I am so far doubtful. Doing this
will definitely make it easier for the oscilloscope to catch since
mine is quite janky and insensitive to trigger edges.