Fix borking reads after performing a write
We don't care what the radio has to say in the case of a write to
its registers, but doing this fixes performing reads after doing
one or more writes. It doesn't hurt to do this anyway either since
another operation cannot occur while a write is happening to the
radio's registers.