Persist MPU regions on process restart
Make sure the profiling regions come back when a process restarts
by calling the initialize and letting it re-initialize even if
there is existing data for the process.

When the process restarts, it does not clear out its entire MPU
configuration. It only replaces the two that it knows existed.
Iterate through the configured regions and clear them out so that
the profiler's old region configurations do not persist through
restarts.