Try shrinking profiling to SP
As an alternative to using the stored faulting address, try moving
the profiling region upper edge down to where the stack pointer is
currently. This is a hack, and I need to investigate why the
address in the MMFAR does not update.

Catch the related memory management faults as new enum variants
for context unstacking and stacking exceptions.

This allows the application to eventually run.