Clear faults to discern recoverability, hint: poor
Clear discovered faults after dealing with them. This means adding
a new function to Chip to allow a caller to specify which fault to
clear.

When encountering stacking or unstacking faults, note that the
process' context is not salvageable, and end the process'
execution. This is always the case, unfortunately.