Loosen lifetime constraints on functions
Though it does not make sense for the PerformanceCounter object to
be around for any lifetime other than the 'static lifetime, it is
more correct to only have the &'static self argument on just the
configure() function. It is necessary there due to the calls to set
itself as a client to other 'static objects.