Make profiling concurrency-safe
Avoid using the MapCell as this causes close instrumentation calls
to induce panics as the stat containers are wholly unavailable if
a single instrumentation point is updating its container. This
change should bring some small performance improvements.

The updated code relies on the assumption that each profiled point
has a unique ID. That is, there is a one-to-one mapping of a Stat
container to a call to account(). Breaking the one-to-one mapping
is undefined behavior.

The new freeze() function should also be reentrant.