Schedule an upcall when TX completes
Send an upcall to the transmitting process once the transmission is
complete. Store the PID of the process the driver is transmitting
for so that we do not have to rely on configured_for for the value,
as, while likely not, that could change unexpectedly.

Move grant-related constants to a submodule for organization.