Fix packet transmission hang on top of receive
Prior to this change, the driver would write to the FIFO and then
switch to transmit mode. When currently listening for a packet
from another source, the radio is in receive mode and switching
from receive mode to transmit mode clears the FIFO, meaning any
ready-to-be-sent packet gets cleared away.

Change to transmit mode before writing to the FIFO.