Make alarms work with batching
If an alarm is set to fire outside the current batching window, we
must open a new batching window immediately after the current one
expires. This is because otherwise, the system will not wake to
execute pending work. Fix this by checking if there is a later
alarm and opening a new window after running all pending syscalls.