Create an observant batching controller
Introduces a new batching controller that simply watches syscall
activity and calculates optimal windows from observed data. It runs
the DBSCAN-like clustering on the data to find an appropriate
time window and returns the window duration from the
find_optimal_window() function.

In order to do the observations, this change effectively undoes the
modifications to the alarm and UART capsules which means the alarm
now directly interacts with timing hardware.

Adds a flush_upcalls function to the batch controller interface to
direct how the kernel should handle upcalls during the batching
phase; this determines whether the kernel will allow them to
execute or not. The ObservantBatching controller directs the kernel
to flush upcalls at every work iteration for proper monitoring of
activity.