Add PrefetchController without prefetch
Get a time-based batching working in PrefetchController while still
allowing the alarm driver to operate the alarm. This batch
controller inspects the state of the alarm hardware to interweave
its own alarms into the mix.

The current hypothesis is that these extra wakeups do not make a
big difference and that we should focus solely on the gains from
parallelizing peripheral operations.