Driver MVP for eInk display
The data/command split requires a fourth pin and that the SPI
transaction not end after sending the command sometimes. Expose
the hold/release_low() functions to allow this.

The display driver can now drive the display to power on and show
what is in the display RAM; currently just noise.